fixed charging for already bought dilation upgrades
The buy dilation upgrade function would subtract the cost before checking if you already had the upgrade, so the NGUdS autobuyer could get stuck in an endless loop of "buying" the same non-rebuyable upgrade over and over again, and clicking on an already bought upgrade would charge for it again. This also caused non-rebuyable upgrades to be charged for twice when bought normally. moving the subtraction after the upgrade code fixes both issues.