Add missing big crunch condition
This fixes an issue introduced in f20c65e07aa8085e1d56c5f2c92da369bcd2f861.
In that commit, the condition that was checking if player can big crunch was removed.
That allowed to big crunch even when it should not be allowed using keyboard shortcut,
and ruin pre-infinity game experience in Grand Runs, NG= or NG- runs.