Fix Time dimension cost being lost from save data upon load
data.cost in save is stored as string, but here we were checking for NaN on data.cost.e,
which does not exists on string.
Fixes a bug in ng-4 where after save reload, the prices were reset to vanilla ones.
Also fixes a bug where the Time Dimension cost from save was lost upon reload, and replaced with vanilla prices.